Removals and additions to make starter branch.
