+++ app/src/main/java/com/example/wordsapp/LetterAdapter.kt
@@ -22,7 +22,6 @@ import android.view.ViewGroup
 import android.view.accessibility.AccessibilityNodeInfo
 import android.widget.Button
 import androidx.annotation.RequiresApi
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 
 /**
@@ -50,9 +49,8 @@ class LetterAdapter :
      */
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): LetterViewHolder {
         val layout = LayoutInflater
-            .from(parent.context)
-            .inflate(R.layout.item_view, parent, false)
-
+                .from(parent.context)
+                .inflate(R.layout.item_view, parent, false)
         // Setup custom accessibility delegate to set the text read
         layout.accessibilityDelegate = Accessibility
         return LetterViewHolder(layout)
@@ -64,16 +62,6 @@ class LetterAdapter :
     override fun onBindViewHolder(holder: LetterViewHolder, position: Int) {
         val item = list.get(position)
         holder.button.text = item.toString()
-
-        // Assigns a [OnClickListener] to the button contained in the [ViewHolder]
-        holder.button.setOnClickListener {
-            // Create an action from WordList to DetailList
-            // using the required arguments
-            val action = WordListFragmentDirections
-                .actionWordListToDetailList(letter = holder.button.text.toString())
-            // Navigate using that action
-            holder.view.findNavController().navigate(action)
-        }
     }
 
     // Setup custom accessibility delegate to set the text read with
