+++ app/src/main/java/com/example/wordsapp/WordAdapter.kt
@@ -16,8 +16,6 @@
 package com.example.wordsapp
 
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
@@ -28,7 +26,7 @@ import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
 
 /**
- * Adapter for the [RecyclerView] in [DetailListFragment].
+ * Adapter for the [RecyclerView] in [DetailActivity].
  */
 class WordAdapter(private val letterId: String, context: Context) :
     RecyclerView.Adapter<WordAdapter.WordViewHolder>() {
@@ -68,6 +66,7 @@ class WordAdapter(private val letterId: String, context: Context) :
 
         // Setup custom accessibility delegate to set the text read
         layout.accessibilityDelegate = Accessibility
+
         return WordViewHolder(layout)
     }
 
@@ -83,14 +82,7 @@ class WordAdapter(private val letterId: String, context: Context) :
         // Set the text of the WordViewHolder
         holder.button.text = item
 
-        // Assigns a [OnClickListener] to the button contained in the [ViewHolder]
-        holder.button.setOnClickListener {
-            val queryUrl: Uri = Uri.parse("${DetailListFragment.SEARCH_PREFIX}${item}")
-            val intent = Intent(Intent.ACTION_VIEW, queryUrl)
-            context.startActivity(intent)
-        }
     }
-
     // Setup custom accessibility delegate to set the text read with
     // an accessibility service
     companion object Accessibility : View.AccessibilityDelegate() {
