+++ app/src/main/res/values/strings.xml
@@ -20,5 +20,4 @@
     <string name="action_switch_layout">Switch Layout</string>
     <string name="look_up_word">Look up word in a Browser Search</string>
     <string name="look_up_words">Show Stored Words</string>
-    <string name="fragment_label">Words That Start With {letter}</string>
 </resources>
