+++ app/src/main/res/values/themes.xml
@@ -22,7 +22,6 @@
         <item name="colorPrimaryVariant">@color/teal_900</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
-        <!-- Unused atm -->
         <item name="colorSecondary">@color/blue_200</item>
         <item name="colorSecondaryVariant">@color/blue_700</item>
         <item name="colorOnSecondary">@color/black</item>
